Allow modifying PodManagementPolicy of a statefulset
It was not possible to update the PodManagementPolicy of
a statefulset.
One use case is for database: while scale up/scale down are typically
done in ordered ready. But in some situation, for example, when several
pods have been deleted it is very important that all pods restart as
quickly as possible in parallel to bring the database cluster back to an
healthy state. Hence being able to change it dynamically.